minishellai/requirements.txt
requests
pyyaml

ruff
